+++ src/main/java/org/hv/biscuits/controller/exception_headler/SqlIntegrityConstraintExceptionHandler.java
@@ -1,6 +1,8 @@
 package org.hv.biscuits.controller.exception_headler;
 
 import org.hv.biscuits.controller.Body;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -14,6 +16,7 @@ import java.sql.SQLIntegrityConstraintViolationException;
  */
 @ControllerAdvice
 public class SqlIntegrityConstraintExceptionHandler {
+    private static final Logger LOGGER = LoggerFactory.getLogger(SqlIntegrityConstraintExceptionHandler.class);
 
     private static final String UNIQUE_KEY = "UK_";
     private static final String DUPLICATE_KEY = "Duplicate";
@@ -23,7 +26,7 @@ public class SqlIntegrityConstraintExceptionHandler {
     @ExceptionHandler(value = SQLIntegrityConstraintViolationException.class)
     @ResponseStatus(HttpStatus.FORBIDDEN)
     public Body handler(SQLIntegrityConstraintViolationException e) {
-        e.printStackTrace();
+        LOGGER.error(e.getMessage());
         if (e.getMessage() == null) {
             return Body.error().title("失败").message(e.toString());
         }