+++ src/main/java/org/hv/biscuits/controller/exception_headler/SystemExceptionHandler.java
@@ -1,6 +1,8 @@
 package org.hv.biscuits.controller.exception_headler;
 
 import org.hv.biscuits.controller.Body;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -12,12 +14,13 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  */
 @ControllerAdvice
 public class SystemExceptionHandler {
+    private static final Logger LOGGER = LoggerFactory.getLogger(SystemExceptionHandler.class);
 
     @ResponseBody
     @ExceptionHandler(value = Exception.class)
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
     public Body handler(Exception e) {
-        e.printStackTrace();
+        LOGGER.error(e.getMessage());
         if (e.getMessage() == null) {
             return Body.error().title("失败").message(e.toString());
         }