+++ src/main/java/org/hv/biscuits/spine/model/Group.java
@@ -23,17 +23,33 @@ public class Group extends AbstractBisEntity {
     private Double sort;
     @Column
     private String additionalServiceId;
+    @Column
+    private String parentUuid;
+    @Column
+    private byte[] img;
+    private String imgBase64;
+
+    private Boolean leafNode;
+
+    public Boolean getLeafNode() {
+        return leafNode;
+    }
+
+    public void setLeafNode(Boolean leafNode) {
+        this.leafNode = leafNode;
+    }
 
     private List<Bundle> bundles;
 
     public Group() {
     }
 
-    public Group(String groupId, String groupName, String serviceId, Double sort) {
+    public Group(String groupId, String groupName, String serviceId, Double sort, String parentUuid) {
         this.groupId = groupId;
         this.groupName = groupName;
         this.serviceId = serviceId;
         this.sort = sort;
+        this.parentUuid = parentUuid;
     }
 
     public String getGroupId() {
@@ -82,5 +98,29 @@ public class Group extends AbstractBisEntity {
 
     public void setAdditionalServiceId(String additionalServiceId) {
         this.additionalServiceId = additionalServiceId;
+    }
+
+    public String getParentUuid() {
+        return parentUuid;
+    }
+
+    public void setParentUuid(String parentUuid) {
+        this.parentUuid = parentUuid;
+    }
+
+    public byte[] getImg() {
+        return img;
+    }
+
+    public void setImg(byte[] img) {
+        this.img = img;
+    }
+
+    public String getImgBase64() {
+        return imgBase64;
+    }
+
+    public void setImgBase64(String imgBase64) {
+        this.imgBase64 = imgBase64;
     }
 }